feat(phase-04-pr04f-m): migrate remaining 8 categories to declarative catalog
Migrates MONITORING, MESSAGING, SECRETS & CONFIG, DNS & CDN, SECURITY & IAM,
CI/CD, DATA & ANALYTICS, and BACKUP to internal/catalog. Adds:
  - internal/catalog/types_{monitoring,messaging,secrets,dns_cdn,security,cicd,data,backup}.go
  - internal/resource/color_{monitoring,messaging,secrets,dns_cdn,security,cicd,data,backup}.go
  - projectRegistry shim for ct-events (ctevent.Project cannot be imported from catalog)

buildResourceTypes() now returns nil — all types come from catalog.
catalog.go allTypes() aggregates all 12 per-category slices.
