+++ internal/resource/color_backup.go
@@ -1,0 +1,5 @@
+package resource
+
+func init() {
+	colorRegistry["backup"] = func(_ Resource) Color { return ColorHealthy }
+}
