--- internal/resource/types_dns_cdn.go
@@ -1,127 +1,0 @@
-package resource
-
-import (
-	"fmt"
-	"strconv"
-)
-
-func dnsCdnResourceTypes() []ResourceTypeDef {
-	return []ResourceTypeDef{
-		{
-			Name:          "Route 53 Hosted Zones",
-			ShortName:     "r53",
-			Aliases:       []string{"r53", "route53", "dns", "hosted-zones"},
-			Category:      "DNS & CDN",
-			CloudTrailKey: "ResourceName:ID",
-			Columns: []Column{
-				{Key: "name", Title: "Name", Width: 36, Sortable: true},
-				{Key: "zone_id", Title: "Zone ID", Width: 30, Sortable: true},
-				{Key: "record_count", Title: "Records", Width: 9, Sortable: true},
-				{Key: "private_zone", Title: "Private", Width: 9, Sortable: true},
-				{Key: "comment", Title: "Comment", Width: 30, Sortable: false},
-			},
-			Color: func(r Resource) Color {
-				s := r.Fields["record_count"]
-				if s != "" {
-					if n, err := strconv.ParseInt(s, 10, 64); err == nil && n <= 2 {
-						return ColorWarning
-					}
-				}
-				return ColorHealthy
-			},
-			Children: []ChildViewDef{{
-				ChildType:      "r53_records",
-				Key:            "enter",
-				ContextKeys:    map[string]string{"zone_id": "ID", "zone_name": "Name"},
-				DisplayNameKey: "zone_name",
-			}},
-		},
-		{
-			Name:          "CloudFront Distributions",
-			ShortName:     "cf",
-			Aliases:       []string{"cf", "cloudfront", "cdn"},
-			Category:      "DNS & CDN",
-			CloudTrailKey: "ResourceName:ID",
-			Columns: []Column{
-				{Key: "domain_name", Title: "Domain Name", Width: 40, Sortable: true},
-				{Key: "distribution_id", Title: "Distribution ID", Width: 16, Sortable: true},
-				{Key: "status", Title: "Status", Width: 12, Sortable: true},
-				{Key: "enabled", Title: "Enabled", Width: 9, Sortable: true},
-				{Key: "aliases", Title: "Aliases", Width: 30, Sortable: false},
-				{Key: "price_class", Title: "Price Class", Width: 16, Sortable: true},
-			},
-			Color: func(r Resource) Color {
-				if r.Fields["enabled"] == "false" {
-					return ColorDim
-				}
-				switch r.Fields["status"] {
-				case "Deployed":
-					return ColorHealthy
-				case "InProgress":
-					return ColorWarning
-				}
-				return ColorHealthy
-			},
-		},
-		{
-			Name:          "ACM Certificates",
-			ShortName:     "acm",
-			Aliases:       []string{"acm", "certificates", "certs"},
-			Category:      "DNS & CDN",
-			CloudTrailKey: "ResourceName:ID",
-			Columns: []Column{
-				{Key: "domain_name", Title: "Domain Name", Width: 40, Sortable: true},
-				{Key: "status", Title: "Status", Width: 14, Sortable: true},
-				{Key: "type", Title: "Type", Width: 14, Sortable: true},
-				{Key: "not_after", Title: "Expires", Width: 22, Sortable: true},
-				{Key: "in_use", Title: "In Use", Width: 8, Sortable: true},
-			},
-			Color: func(r Resource) Color {
-				switch r.Fields["status"] {
-				case "ISSUED":
-					dl := r.Fields["days_left"]
-					if dl == "expired" {
-						return ColorBroken
-					}
-					if dl != "" {
-						var n int
-						if _, err := fmt.Sscanf(dl, "%d days", &n); err == nil {
-							if n < 7 {
-								return ColorBroken
-							}
-							if n < 30 {
-								return ColorWarning
-							}
-						}
-					}
-					if r.Fields["in_use"] == "false" {
-						return ColorWarning
-					}
-					return ColorHealthy
-				case "PENDING_VALIDATION":
-					return ColorWarning
-				case "EXPIRED", "REVOKED", "FAILED", "VALIDATION_TIMED_OUT":
-					return ColorBroken
-				case "INACTIVE":
-					return ColorDim
-				}
-				return ColorHealthy
-			},
-		},
-		{
-			Name:          "API Gateways",
-			ShortName:     "apigw",
-			Aliases:       []string{"apigw", "apigateway", "api-gateway"},
-			Category:      "DNS & CDN",
-			CloudTrailKey: "ResourceName:ID",
-			Columns: []Column{
-				{Key: "name", Title: "Name", Width: 28, Sortable: true},
-				{Key: "api_id", Title: "API ID", Width: 14, Sortable: true},
-				{Key: "protocol", Title: "Protocol", Width: 12, Sortable: true},
-				{Key: "endpoint", Title: "Endpoint", Width: 50, Sortable: false},
-				{Key: "description", Title: "Description", Width: 30, Sortable: false},
-			},
-			Color: func(_ Resource) Color { return ColorHealthy },
-		},
-	}
-}
